+++ pyssata/lib/pyr_compute_slopes.py
@@ -1,7 +1,8 @@
 import numpy as np
-from pyssata import gpuEnabled
+
 from pyssata import xp
 from pyssata import cpuArray
+from pyssata import standard_dtype
 
 def pyr_compute_slopes(frame, ind_pup, SHLIKE=False, INTENSITY_BASED=False, norm_fact=None, threshold=None):
     """
@@ -25,10 +26,10 @@ def pyr_compute_slopes(frame, ind_pup, SHLIKE=False, INTENSITY_BASED=False, norm
         raise ValueError('INTENSITY_BASED and SHLIKE keywords cannot be set together.')
 
     # Extract intensity arrays for each sub-pupil
-    intensity = xp.array( [cpuArray(frame).flat[cpuArray(ind_pup)[:, i]].reshape(-1) for i in range(4)] )
+    intensity = xp.array( [cpuArray(frame).flat[cpuArray(ind_pup)[:, i]].reshape(-1) for i in range(4)], dtype=standard_dtype )
 
     # Compute total intensity
-    flux = xp.sum(xp.array([xp.sum(arr) for arr in intensity]))
+    flux = xp.sum(xp.array([xp.sum(arr) for arr in intensity], dtype=standard_dtype))
     
     if threshold is not None:
         # Apply thresholding
@@ -36,7 +37,7 @@ def pyr_compute_slopes(frame, ind_pup, SHLIKE=False, INTENSITY_BASED=False, norm
     
     total_intensity = np.sum([np.sum(cpuArray(arr)) for arr in intensity])
 
-    total_intensity = xp.array(total_intensity)
+    total_intensity = xp.array(total_intensity, dtype=standard_dtype)
 
     n_subap = ind_pup.shape[0]
 
@@ -51,7 +52,7 @@ def pyr_compute_slopes(frame, ind_pup, SHLIKE=False, INTENSITY_BASED=False, norm
             if not SHLIKE:
                 factor = n_subap / total_intensity
             else:
-                inv_factor = xp.array([xp.sum(arr) for arr in intensity])
+                inv_factor = xp.array([xp.sum(arr) for arr in intensity], dtype=self.dtype)
                 inv_factor[inv_factor <= 0] = 1e-6
                 factor = 1.0 / inv_factor
                 factor[inv_factor <= 0] = 0.0
@@ -60,10 +61,10 @@ def pyr_compute_slopes(frame, ind_pup, SHLIKE=False, INTENSITY_BASED=False, norm
             sy = (intensity[1] + intensity[2] - intensity[3] - intensity[0]) * factor
     else:
         if INTENSITY_BASED:
-            sx = xp.zeros(2 * n_subap)
-            sy = xp.zeros(2 * n_subap)
+            sx = xp.zeros(2 * n_subap, dtype=self.dtype)
+            sy = xp.zeros(2 * n_subap, dtype=self.dtype)
         else:
-            sx = xp.zeros(n_subap)
-            sy = xp.zeros(n_subap)
+            sx = xp.zeros(n_subap, dtype=self.dtype)
+            sy = xp.zeros(n_subap, dtype=self.dtype)
 
     return sx, sy, flux
